split pumpchart and psychart tests
