--- packages/psychart/src/both.test.ts
@@ -1,158 +1,0 @@
-import * as T6 from 't6';
-import * as SMath from 'smath';
-import { Psychart, Pumpchart } from './index.js';
-import { defaultOptions, regions } from './psychart/defaults.js';
-import { PsyState } from './psychart/psystate.js';
-import { f, zero } from './pumpchart/lib.js';
-import { FlowUnits } from './pumpchart/units.js';
-
-T6.eq(Psychart.getRegionNamesAndTips().length, Object.entries(regions).length);
-T6.eq(Pumpchart.getFlowUnits().length, Object.entries(FlowUnits).length);
-
-{
-    let caught: boolean;
-
-    caught = false;
-    try {
-        new PsyState({ db: 70, other: 1.60, measurement: 'dbrh' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught RH was out of bounds.');
-
-    caught = false;
-    try {
-        new PsyState({ db: 70, other: -0.60, measurement: 'dbrh' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught RH was out of bounds.');
-
-    caught = false;
-    try {
-        new PsyState({ db: 70, other: 80, measurement: 'dbwb' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught wb > db.');
-
-    caught = false;
-    try {
-        new PsyState({ db: 60, other: -10e-3, measurement: 'dbhr' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught invalid (low) humidity ratio');
-
-    caught = false;
-    try {
-        new PsyState({ db: 60, other: 20e-3, measurement: 'dbhr' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught invalid (high) humidity ratio');
-
-    caught = false;
-    try {
-        new PsyState({ db: 60, other: 10e-3, measurement: 'dbhr' }, defaultOptions);
-    } catch (e) {
-        console.log(e);
-        caught = true;
-    }
-    T6.isFalse(caught, 'Caught valid humidity ratio');
-
-    caught = false;
-    try {
-        new PsyState({ db: 60, other: 10, measurement: 'dbh' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught invalid (low) enthalpy');
-
-    caught = false;
-    try {
-        new PsyState({ db: 60, other: 30, measurement: 'dbh' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isTrue(caught, 'Uncaught invalid (high) enthalpy');
-
-    caught = false;
-    try {
-        new PsyState({ db: 60, other: 20, measurement: 'dbh' }, defaultOptions);
-    } catch {
-        caught = true;
-    }
-    T6.isFalse(caught, 'Caught valid enthalpy');
-
-    // We will not be testing the accuracy of psychrolib, but will make sure that PsyState computes.
-    const ps1: PsyState = new PsyState({ db: 70, other: 0.50, measurement: 'dbrh' }, defaultOptions);
-    T6.eq(ps1.db, 70);
-    T6.eq(ps1.rh, 0.50);
-    // Just make sure that values are greater than zero.
-    T6.gt(ps1.dp, 0);
-    T6.gt(ps1.h, 0);
-    T6.gt(ps1.hr, 0);
-    T6.gt(ps1.s, 0);
-    T6.gt(ps1.v, 0);
-    T6.gt(ps1.vp, 0);
-    T6.gt(ps1.wb, 0);
-    const xy = ps1.toXY();
-    T6.gt(xy.x, 0);
-    T6.gt(xy.y, 0);
-}
-
-{
-    const f1: f = x => x * x - 9; // Has 2 solutions
-    T6.isTrue(SMath.approx(zero(f1, 0, 10), 3));
-    T6.isTrue(SMath.approx(zero(f1, 0, -10), -3));
-    const f2: f = x => x * x * x - 64;
-    T6.isTrue(SMath.approx(zero(f2, 0, 10), 4));
-    T6.isTrue(SMath.approx(zero(f2, 0, 4), 4));
-    T6.isTrue(SMath.approx(zero(f2, 4, 0), 4));
-    let caught: boolean;
-    let message: string;
-    caught = false;
-    message = '';
-    try {
-        zero(f1, -5, -4);
-    } catch (e) {
-        caught = true;
-        message = (e as Error).message;
-    }
-    T6.isTrue(caught, 'Did not catch f1 [-5, -4]');
-    T6.is(message, 'A solution cannot be found.');
-    caught = false;
-    message = '';
-    try {
-        zero(f1, -5, 5); // Cannot find both solutions at once
-    } catch (e) {
-        caught = true;
-        message = (e as Error).message;
-    }
-    T6.isTrue(caught, 'Did not catch f1 [-5, 5]');
-    T6.is(message, 'A solution cannot be found.');
-    caught = false;
-    message = '';
-    try {
-        zero(f2, -5, -10);
-    } catch (e) {
-        caught = true;
-        message = (e as Error).message;
-    }
-    T6.isTrue(caught, 'Did not catch f2');
-    T6.is(message, 'A solution cannot be found.');
-    T6.isTrue(SMath.approx(zero(Math.sin, 1, 5), Math.PI), 'sin');
-    T6.isTrue(SMath.approx(zero(Math.cos, 0, 4), Math.PI / 2), 'cos');
-    T6.isTrue(SMath.approx(zero(Math.log, 0, 10), 1), 'log');
-    caught = false;
-    message = '';
-    try {
-        zero(Math.log, -1, 10); // Bad domain
-    } catch (e) {
-        caught = true;
-        message = (e as Error).message;
-    }
-    T6.isTrue(caught, 'Did not catch log(-1)');
-    T6.is(message, 'f(-1) is NaN');
-}
